% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_choices.R
\name{get_choices}
\alias{get_choices}
\title{Get Choices for Each Round}
\usage{
get_choices(phase_data)
}
\arguments{
\item{phase_data}{One of the three phase objects defined by calling \code{\link{get_phases}}.}
}
\value{
Returns a dataframe with five columns:
\itemize{
\item \emph{option:} The option the subject chose.
This should correspond to the icon order as presented in the task.
\item \emph{choice:} The icon the subject chose.
\item \emph{group:} The (payment weighting) group the icon corresponds to.
\item \emph{phase:} The phase the choices were made in (constant for each \code{phase} object).
\item \emph{round:} The round number these choices were made in, specific to the phase.
}
}
\description{
This function gets the choices for each round for each phase.
}
\examples{
\dontrun{get_choices(phase$two)}
}
