% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_groups.R
\name{get_group_info}
\alias{get_group_info}
\title{Get Group Reward and Probability Info for All Phases}
\usage{
get_group_info(data)
}
\arguments{
\item{data}{The data output from \code{\link{read_prpt}}}
}
\value{
A dataframe with 15 rows and seven columns. It has the reward amount
  and probability of receiving a reward as a function of each image's group
  per phase. This also has the icon to group mapping for the subject. Most
  importantly, it has the EU and rank columns. These are used to determine
  optimal choice when measuring behavioral performance. This can be used in
  conjunction with the output of \code{\link{get_choices}} or
  \code{\link{process_choice_data}}.
}
\description{
Get Group Reward and Probability Info for All Phases
}
