% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_groups.R
\name{get_group_reward_probs}
\alias{get_group_reward_probs}
\title{Get One Phase's Group Reward and Probability Info}
\usage{
get_group_reward_probs(phase_data)
}
\arguments{
\item{phase_data}{One of the three phase objects defined by calling
\code{\link{get_phases}}.}
}
\value{
A dataframe with four columns. It has the reward and probability of
  receiving it as a function of each image's group per phase. This can be
  compared to the output of \code{\link{get_choices}} or
  \code{\link{process_choice_data}}. This function is called by
  \code{\link{get_group_info}} which combines the information for all three of
  the phases.
}
\description{
Get One Phase's Group Reward and Probability Info
}
