% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_icons.R
\name{get_icons}
\alias{get_icons}
\title{Get Icons for Each Round}
\usage{
get_icons(data, specific_phase = NULL)
}
\arguments{
\item{data}{The data output from \code{\link{read_prpt}}}

\item{specific_phase}{If this argument is specified, the icons and groups for
a specific phase will be returned. If this is not specified, the icons and
groups for all phases will be returned.}
}
\value{
Returns a dataframe with ten columns: \itemize{ \item \emph{phase:}
  The phase these icons were displayed in (constant for each \code{phase}
  object). \item \emph{round:} The round number these icons were displayed
  in, specific to the phase. \item \emph{icon1:} The icon presented in the
  first order position. \item \emph{icon1Group:} The group mapping for the
  icon presented in the first order position. \item \emph{icon1Prob:} The
  probability of getting a reinforcing image by choosing icon 1. \item
  \emph{icon1Sign:} The relative sign of the reinforcing image (positive or
  negative) by choosing icon 1. \item \emph{icon2:} The icon presented in the
  second order position. \item \emph{icon2Group:} The group mapping for the
  icon presented in the second order position. \item \emph{icon2Prob:} The
  probability of getting a reinforcing image by choosing icon 2. \item
  \emph{icon2Sign:} The relative sign of the reinforcing image (positive or
  negative) by choosing icon 2. }
}
\description{
This function parses the information for the icons presented in each round.
}
\examples{
\dontrun{get_icons(sampleChoiceData, specific_phase = 1)}
}
