% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_prpt.R
\name{get_phase_icons}
\alias{get_phase_icons}
\title{Get Icons for Each Round per Phase}
\usage{
get_phase_icons(phase_data)
}
\arguments{
\item{phase_data}{One of the three phase objects defined by calling
\code{\link{getPhases}}.}
}
\value{
This returns a dataframe with four columns: \itemize{\item
  \emph{phase:} The phase these icons were displayed in (constant for each
  \code{phase} object). \item \emph{round:} The round number these icons were
  displayed in, specific to the phase. \item \emph{icon1:} The icon presented
  in the first order position. \item \emph{icon2:} The icon presented in the
  second order position. }
}
\description{
This function gets the icons for each round in each phase.
}
