% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_prpt.R
\name{get_phases}
\alias{get_phases}
\title{Get Task Phases}
\usage{
get_phases(data)
}
\arguments{
\item{data}{The data output from \code{\link{read_prpt}}}
}
\description{
This function takes the data column from the output of
\code{\link{read_prpt}} and separates the three phases of the task, saving
them to the environment \code{phrase}, which is initialized behind the
scenes. It also separates the initial icon rating phase and the final icon
rating phase. This function \strong{must} be run in order to use the other
functions that parse the choice data. If it is not, the phases will need to
be isolated by hand and assigned to the phase environment (or local/global
variables). See the \emph{Details} section for more on how to access the three
phases.
}
\details{
The task and icon rating phases may be accessed by calling the variables:
\itemize{ \item \code{phase$one} \item \code{phase$two} \item
\code{phase$three} \item \code{phase$forced} \item \code{phase$finalIcon}}
}
