% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_reinforcers.R
\name{get_reinforcers}
\alias{get_reinforcers}
\title{Get Reinforcer Information}
\usage{
get_reinforcers(phase_data)
}
\arguments{
\item{phase_data}{One of the three phase objects defined by calling
\code{\link{get_phases}}.}
}
\value{
A data frame with four columns:
* \emph{phase:} The phase the choices were made in.
* \emph{round number:} The round number, specific to the phase.
* \emph{reward:} The reward (in dollars) participants received from their choice.
* \emph{cumulative_reward:} The cumulative reward since the beginning of the game (phase 1, round 1).
}
\description{
This function gets the monetary reinforcement (reward) shown at each round.
}
