% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{read_prpt}
\alias{read_prpt}
\title{Read in data from PRPT task}
\usage{
read_prpt(file)
}
\arguments{
\item{file}{A character string that contains the path to the file.}
}
\value{
A data frame that contains the PRPT behavioral data with time stamps.
}
\description{
Read in data from PRPT task
}
