% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_data}
\alias{sample_data}
\title{Sample Data from the PRPT Task}
\format{
A data frame with 4537 rows and 2 columns
\describe{
  \item{time}{Time stamps from PRPT task.}
  \item{data}{data from PRPT task.}
}
}
\source{
Kishida Lab (researcher)
}
\usage{
sample_data
}
\description{
A dataset containing output from the PRPT task performed by
a study team member. This object has been read in using the
\code{\link{read_prpt}} function.
}
\keyword{datasets}
