% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_phases.R
\name{get_phases}
\alias{get_phases}
\title{Get Task Phases}
\usage{
get_phases(data, final_icon_rating = FALSE)
}
\arguments{
\item{data}{The data output from \code{\link{read_prpt}}}

\item{final_icon_rating}{LOGICAL: FALSE by default for the majority of PRPT
tasks that don't ask for final icon ratings. TRUE, and the phases will
return those trials}
}
\value{
A list of data for each phase in the PRPT task.
}
\description{
This function takes the data column from the output of
\code{\link{read_prpt}} and separates the three phases of the task, saving
them to the environment \code{phrase}, which is initialized behind the
scenes. It also separates the initial icon rating phase and the final icon
rating phase. This function \strong{must} be run in order to use the other
functions that parse the choice data. If it is not, the phases will need to
be isolated by hand and assigned to the phase environment (or local/global
variables). See the \emph{Details} section for more on how to access the
three phases.
}
\examples{

phase_info <- get_phases(sample_data)

# Print different phase information by indexing into the object.
# Phase three can be accessed as follows, for example:
phase_info$phaseThree

}
