% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_prpt.R
\name{get_reinforcers}
\alias{get_reinforcers}
\title{Get Reinforcer Information}
\usage{
get_reinforcers(data)
}
\arguments{
\item{data}{The data output from \link{read_prpt}.}
}
\value{
A data frame with three rows: round number, reward (monetary value),
  cumulative  reward.
}
\description{
This function gets the monetary reinforcement (reward) shown at each round.
}
